refactor(check): keep stale-install probe worktree-local
