+++ test/integration/provider-scenarios/doctor.test.ts
@@ -1,8 +1,7 @@
 import assert from 'node:assert/strict';
 import fs from 'node:fs';
 import http from 'node:http';
-import path from 'node:path';
-import { afterEach, test, vi } from 'vitest';
+import { test } from 'vitest';
 import type { AndroidAdbProvider } from '../../../src/platforms/android/adb-executor.ts';
 import { assertRpcOk } from './assertions.ts';
 import {
@@ -17,43 +16,6 @@ import {
   withProviderScenarioResource,
   withProviderScenarioTempDir,
 } from './harness.ts';
-
-// The node-modules probe reads the root that `findProjectRoot()` resolves to. Left alone
-// it is this source checkout, which is what the source-context assertion below wants; the
-// packaged assertion points it at a directory shaped like a published install instead, so
-// the real detection runs against a real packaged filesystem rather than a stubbed verdict.
-const projectRootState = vi.hoisted(() => ({ override: undefined as string | undefined }));
-
-vi.mock('../../../src/utils/version.ts', async () => {
-  const actual = await vi.importActual<typeof import('../../../src/utils/version.ts')>(
-    '../../../src/utils/version.ts',
-  );
-  return {
-    ...actual,
-    findProjectRoot: () => projectRootState.override ?? actual.findProjectRoot(),
-  };
-});
-
-afterEach(() => {
-  projectRootState.override = undefined;
-});
-
-/** A published agent-device on disk: bin/ and dist/, no pnpm-lock.yaml, no node_modules. */
-function layOutPackagedInstall(root: string): string {
-  const packageRoot = path.join(root, 'node_modules', 'agent-device');
-  fs.mkdirSync(path.join(packageRoot, 'dist', 'src'), { recursive: true });
-  fs.mkdirSync(path.join(packageRoot, 'bin'), { recursive: true });
-  fs.writeFileSync(
-    path.join(packageRoot, 'package.json'),
-    `${JSON.stringify({ name: 'agent-device', version: '0.0.0-test' })}\n`,
-  );
-  assert.equal(
-    fs.existsSync(path.join(packageRoot, 'pnpm-lock.yaml')),
-    false,
-    'sanity: a packaged install ships no lockfile',
-  );
-  return packageRoot;
-}
 
 test('Provider-backed integration doctor infers Android RN/Metro readiness through daemon route without resolving a default device', async () => {
   const server = await startMetroStatusServer();
@@ -324,77 +286,6 @@ test('Provider-backed integration doctor surfaces a platform inventory failure e
       assertDoctorCheck(data, 'device', 'pass');
       const failure = assertDoctorCheck(data, 'device-apple', 'warn');
       assert.match(failure.summary, /simctl/);
-    },
-  );
-});
-
-test('Provider-backed integration doctor reports the node-modules probe in a source checkout', async () => {
-  // The #1963 value, asserted at the route: from a real source worktree the probe is
-  // present and passing. Without this, the two omission assertions below could both hold
-  // for the trivial reason that the probe never runs anywhere.
-  await withProviderScenarioResource(
-    async () =>
-      await createProviderScenarioHarness({
-        deviceInventoryProvider: async () => [PROVIDER_SCENARIO_ANDROID],
-      }),
-    async (daemon) => {
-      const response = await daemon.callCommand('doctor', [], { platform: 'android' });
-      assertRpcOk(response);
-      const data = response.json.result.data;
-      assertDoctorCheck(data, 'node-modules', 'pass');
-    },
-  );
-});
-
-test('Provider-backed integration doctor omits the node-modules probe for a packaged install', async () => {
-  // A published agent-device ships neither pnpm-lock.yaml nor an installed snapshot, so an
-  // unconditional probe gave every end user a spurious node-modules line and dragged the
-  // overall status to warn. The probe must be absent, and the status must not be degraded.
-  await withProviderScenarioTempDir(
-    'agent-device-doctor-packaged-',
-    async (root) =>
-      await withProviderScenarioResource(
-        async () =>
-          await createProviderScenarioHarness({
-            deviceInventoryProvider: async () => [PROVIDER_SCENARIO_ANDROID],
-          }),
-        async (daemon) => {
-          projectRootState.override = layOutPackagedInstall(root);
-
-          const response = await daemon.callCommand('doctor', [], { platform: 'android' });
-          assertRpcOk(response);
-          const data = response.json.result.data;
-          assertNoDoctorCheck(data, 'node-modules');
-          assert.equal(data.status, 'pass', JSON.stringify(data.checks));
-          assert.equal(
-            data.checks.some((check: { summary: string }) => /lockfile/i.test(check.summary)),
-            false,
-            `no packaged check may mention a lockfile: ${JSON.stringify(data.checks)}`,
-          );
-        },
-      ),
-  );
-});
-
-test('Provider-backed integration doctor --remote omits the node-modules probe', async () => {
-  // --remote diagnoses the caller's connection to a remote daemon; the daemon's own root
-  // would describe the server deployment rather than the caller's worktree, so the probe
-  // cannot answer #1963 there. The project root is left as this real source checkout, so a
-  // missing gate would surface the check rather than silently agreeing with the assertion.
-  await withProviderScenarioResource(
-    async () =>
-      await createProviderScenarioHarness({
-        deviceInventoryProvider: async () => [PROVIDER_SCENARIO_ANDROID],
-      }),
-    async (daemon) => {
-      const response = await daemon.callCommand('doctor', [], {
-        remote: true,
-        daemonBaseUrl: 'https://example.invalid/agent-device',
-      });
-      assertRpcOk(response);
-      const data = response.json.result.data;
-      assert.equal(data.status, 'pass', JSON.stringify(data.checks));
-      assertNoDoctorCheck(data, 'node-modules');
     },
   );
 });
